Create Stellar Mission - version 0.12
Title screen takes you to the map screen, load game isn't implemented yet, credits takes you to credits "room" and needs graphics. 

The government building button at the bottom of the map screen is an example of how I think we should implement the building buttons (don't click it right now, right now it exits the game as a test that the button works). The object "generic_button" is an object that is implemented for the title screen buttons.